feat: Icons paths fixed. Some refactoring. Ambassador page complete layout
