+++ src/components/Layout/CheckBoxes/CheckBoxes.tsx
@@ -27,7 +27,7 @@ const BpIcon = styled('span')(() => ({
     width: `${vars.sizeIcon}`,
     height: `${vars.sizeIcon}`,
     backgroundImage:
-      "url(\"constants:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath" +
+      "url(\"data:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath" +
       " fill-rule='evenodd' clip-rule='evenodd' d='M12 5c-.28 0-.53.11-.71.29L7 9.59l-2.29-2.3a1.003 " +
       "1.003 0 00-1.42 1.42l3 3c.18.18.43.29.71.29s.53-.11.71-.29l5-5A1.003 1.003 0 0012 5z' fill='%2380818F'/%3E%3C/svg%3E\")",
     content: '""',
@@ -44,7 +44,7 @@ const BpCheckedIcon = styled(BpIcon)({
     width: `${vars.sizeIcon}`,
     height: `${vars.sizeIcon}`,
     backgroundImage:
-      "url(\"constants:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath" +
+      "url(\"data:image/svg+xml;charset=utf-8,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Cpath" +
       " fill-rule='evenodd' clip-rule='evenodd' d='M12 5c-.28 0-.53.11-.71.29L7 9.59l-2.29-2.3a1.003 " +
       "1.003 0 00-1.42 1.42l3 3c.18.18.43.29.71.29s.53-.11.71-.29l5-5A1.003 1.003 0 0012 5z' fill='%231A1A1A'/%3E%3C/svg%3E\")",
     content: '""',
