+++ src/pages/Home.tsx
@@ -1,17 +1,5 @@
-import { ambassadorColumns } from '@constants/tableColumns';
-import { ambassadorRows } from '@constants/tableRows';
-import Table from '@/components/Table/Table';
-import Content from '@/components/Layout/Content/Content';
-
 const Home = () => {
-  return (
-    <>
-      <div style={{ gridArea: 'controls', height: '76px' }} />
-      <Content>
-        <Table columns={ambassadorColumns} rows={ambassadorRows} />
-      </Content>
-    </>
-  );
+  return <div />;
 };
 
 export default Home;
